DEX-560 gRPC issues
Review fixes
