+++ dex-it-common/src/main/scala/com/wavesplatform/dex/it/docker/BaseContainer.scala
@@ -12,7 +12,7 @@ import com.github.dockerjava.core.command.ExecStartResultCallback
 import com.typesafe.config.Config
 import com.wavesplatform.dex.it.api.HasWaitReady
 import com.wavesplatform.dex.it.cache.CachedData
-import com.wavesplatform.dex.it.config.Implicits.ConfigOps
+import com.wavesplatform.dex.settings.utils.ConfigOps.ConfigOps
 import com.wavesplatform.utils.ScorexLogging
 import org.testcontainers.images.builder.Transferable
